// WADNR.Web/src/app/shared/generated/model/organization-type-lookup-item.ts
/**
 * WADNR.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class OrganizationTypeLookupItem { 
    OrganizationTypeID?: number;
    OrganizationTypeName?: string | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface OrganizationTypeLookupItemForm { 
    OrganizationTypeID?: FormControl<number>;
    OrganizationTypeName?: FormControl<string>;
}

export class OrganizationTypeLookupItemFormControls { 
    public static OrganizationTypeID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static OrganizationTypeName = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
}